requirements.txt
nltk
wikipedia
sklearn
google
colorlog
click
geoip2
python-whois
python-dateutil
requests
docker-compose
pywebcopy
pyquery
w3lib
parse
psycopg2
ijson
aiml
fastapi
uvicorn
pandas
parameterized
tabulate
docker-py
httpie
pytest
pytest-sugar
pytest-cov
pytest-instafail
pytest-picked
pytest-tldr
bandit
pyaudio
SpeechRecognition
pocketsphinx
google-api-python-client
oauth2client
gTTS
newspaper3k